changed the files name and removed unnecessary files
